update menu after last item
Runs loop() once the last menu item is selected and the user presses another time *. This offers the advantage that the wifi/battery status get re-checked and updated along the the way. In case the wifi connection gets lost/re-connected, it would automatically show/hide the net based menu item.