Add device selection logic to TFT_eSPI/User_Setup_Select.h
Including a file from the Arduino sketch using an absolute path
seems to be the best/cleanest way to do this, although it's not super elegant.

The alternative is using symlinks/hardlinks/copies to select the device,
but that would be a deviation from the way the device is selected
in the Arduino sketch, and would mean the user has to change the library
every time they compile for a new device. So that seems worse.